Move access-json logging related classes to module dropwizard-request-logging-logback-access
