+++ dropwizard-json-logging/src/main/resources/META-INF/services/io.dropwizard.logging.common.layout.DiscoverableLayoutFactory
@@ -1,2 +1,1 @@
-io.dropwizard.logging.json.AccessJsonLayoutBaseFactory
 io.dropwizard.logging.json.EventJsonLayoutBaseFactory
