+++ dropwizard-request-logging-logback-access/src/main/java/io/dropwizard/request/logging/logback/access/json/AccessAttribute.java
@@ -1,4 +1,4 @@
-package io.dropwizard.logging.json;
+package io.dropwizard.request.logging.logback.access.json;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
