+++ dropwizard-request-logging-logback-access/src/main/resources/META-INF/services/io.dropwizard.logging.common.layout.DiscoverableLayoutFactory
@@ -1,0 +1,1 @@
+io.dropwizard.request.logging.logback.access.json.AccessJsonLayoutBaseFactory
